services/dashboard-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
kafka-python
python-multipart
python-jose[cryptography]
passlib[bcrypt]
structlog
python-json-logger
httpx
